+++ example/GetSocialSdkDemo/Assets/Facebook/Editor/FacebookPostprocess.cs
@@ -18,6 +18,8 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
+using GetSocialSdk.Editor;
+
 namespace UnityEditor.FacebookEditor
 {
     using System;
@@ -51,7 +53,7 @@ namespace UnityEditor.FacebookEditor
             if (target == BuildTarget.Android)
             {
                 // The default Bundle Identifier for Unity does magical things that causes bad stuff to happen
-                if (PlayerSettings.bundleIdentifier == "com.Company.ProductName")
+                if (PlayerSettingsCompat.bundleIdentifier == "com.Company.ProductName")
                 {
                     Debug.LogError("The default Unity Bundle Identifier (com.Company.ProductName) will not work correctly.");
                 }