+++ example/GetSocialSdkDemo/Assets/Plugins/iOS/GetSocial/GetSocialUI.framework/Headers/GetSocialUIView.h
@@ -21,6 +21,13 @@
 /** @name Methods */
 
 /*!
+ * @abstract Sets listeners to handle GetSocial view state events.
+ * @param onOpen Block is called when GetSocial view is opened.
+ * @param onClose Block is called when GetSocial view is closed.
+ */
+- (void)setHandlerForViewOpen:(void (^)())onOpen close:(void (^)())onClose;
+
+/*!
  *  @abstract Shows GetSocialUIView.
  *
  *  @result YES, if view is presented without any error, otherwise NO.